requirements.txt
openai
python-dotenv
pypdf
pdf2image
more_itertools
chromadb